% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.0.data_documentation.R
\docType{data}
\name{sorghum_compendium}
\alias{sorghum_compendium}
\title{Sorghum RNAseq Compendium}
\format{
A list containing multiple RNAseq datasets.
}
\source{
Internal TAMU RNAseq compendium
}
\usage{
sorghum_compendium
}
\description{
A compiled RNAseq dataset for Sorghum bicolor.
}
\keyword{datasets}
